+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Change log
 - Added some code to embed videos from YouTube and Vimeo without
   requiring oEmbed.
 - Dropped compatibility with Python 3.5.
+- Made :func:`~feincms3.renderer.render_in_context` create its own
+  ``Context`` if the context passed is ``None``.
 
 
 `0.36`_ (2020-08-07)